check for allocation failures from calloc
the cgo documentation states that calling C.malloc
actually calls into a generated wrapper that checks for
nil pointer returns and crashes the program if the
allocation failed, or allocates with size 1 if the
requested allocation was zero sized. in this way, it
guarantees that the returned result is always non-nil.
this defines and uses a calloc wrapper that provides
the same behavior.

Change-Id: I7868c522979317ef0cac10cd5177b8b7a7734bf2