GNUmakefile: add automation for adding assets to releases
It can be difficult for non-Go developers to understand how to build
and use uplink-c.  To make things easier for others, we should supply
precompiled shared object binaries for the most common use cases.

This change adds make automation to ease the developerment burden of
creating cross compiled release assests for linux, macos, and windows.