scripts: Use realpath to suppress double path separator
The script concatenates the output path, passed as command line
argument, with the file names. If the parsed argument ends with the path
separator (the current Makefile does it) the output file path contained
two path separators.

The script worked without this change, but it's more correct to
normalize the path, and realpath does it.

Change-Id: Icf263844d3c9980063574763c45ad50cbb67fd71